Make a type-gen directory
- refactor: extract generators for each type
- refactor: move generators into separate file
- refactor: remove unnecessary class FileModule
- refactor: keep all type-gen concerns together
