+++ src/jsonix.d.ts
